mp4: Extract attachments upon first request
Instead of each time DEMUX_GET_ATTACHMENT gets invoked

+++ modules/demux/mp4/attachments.h
@@ -21,7 +21,7 @@
 #ifndef VLC_MP4_ATTACHMENTS_H_
 #define VLC_MP4_ATTACHMENTS_H_
 
-int MP4_GetAttachments( const MP4_Box_t *, input_attachment_t *** );
+size_t MP4_GetAttachments( const MP4_Box_t *, input_attachment_t *** );
 const MP4_Box_t *MP4_GetMetaRoot( const MP4_Box_t *, const char ** );
 int MP4_GetCoverMetaURI( const MP4_Box_t *,  const MP4_Box_t *,
                          const char *, vlc_meta_t * );
